requirements.txt
beautifulsoup4==4.8.2
blackboxprotobuf
packaging==20.1
protobuf==3.10.0
PyCryptodome
PySimpleGUI==4.16.0
simplekml
xmltodict
python-magic-bin==0.4.14
xlrd==1.2.0
PyMuPDF
openpyxl